Change handle hierarchy to not use inheritance.
Having the more specific handles inherit from ManagedHandle caused
a problem with arrays of handles.  Inheritence changes the memory
layout, adding a member, so a seq[Instance] has a different layout
than a seq of the raw OSPObject handle.  It's common to pass arrays of
handles in to OSPRay, so the scoped way with ManagedHandle can not be
passed as shared data.

Trying to work around this by storing an array of raw OSPObject handles
instead works, but opens you up to having to manually call retain and
release for the handles again, negating the point of having
ManagedHandle.

Changed the implementation so instead of inheriting, the more specific
objects include a ManagedHandle member. With a little template magic
to paper over the structure differences, and add a converter that can
decay a more specific instance to a ManagedHandle as a replacement for
inheritance, it works well.